// admin/src/client/types.gen.ts
export type Source = {
    /**
     * Id
     */
    id: string;
    /**
     * Name
     */
    name: string;
    /**
     * Endpoint
     */
    endpoint?: string | null;
    /**
     * Schedule
     */
    schedule?: {
        [key: string]: unknown;
    } | null;
    /**
     * Beat
     */
    beat?: string | null;
    /**
     * Geo
     */
    geo?: string | null;
    /**
     * Related Sources
     */
    related_sources?: Array<string> | null;
    /**
     * Registered At
     */
    registered_at?: string | null;
    /**
     * Health
     */
    health?: {
        [key: string]: unknown;
    } | null;
    source_tier?: SourceTier;
    /**
     * Outlet
     */
    outlet?: string | null;
    status?: SourceStatus;
};
// ...
export type SourceStatus = 'active' | 'inactive';
// ...
export type SourceTier = 'primary' | 'secondary';